20221203_binance_opt_position/requirements.txt
websocket-client>=1.3.3